refactor: ammend file names and structure of Slice
