+++ packages/slice/__tests__/web/responsive.web.test.js
@@ -5,7 +5,7 @@ import renderer from "react-test-renderer";
 import {
   SliceContainer,
   getSeparator
-} from "../../templates/shared.responsive";
+} from "../../templates/styles/shared.responsive";
 import {
   ChildContainer,
   getChildrenContainer
