+++ packages/slice/slice.js
@@ -1,6 +1,6 @@
 import StandardSlice from "./templates/standard";
-import standardRoles from "./templates/default/roles";
+import standardRoles from "./templates/standard/roles";
 import LeadAndTwoSlice from "./templates/leadandtwo";
-import leadAndTwoRoles from "./templates/lead/roles";
+import leadAndTwoRoles from "./templates/leadandtwo/roles";
 
 export { StandardSlice, standardRoles, LeadAndTwoSlice, leadAndTwoRoles };
