+++ packages/slice/templates/standard/roles.js
