+++ packages/slice/templates/styles/index.js
