+++ packages/slice/templates/styles/shared.responsive.web.js
